IO protection and RGB led integration
- GPIO helper : io protection
- Mq3 : get lower value
- Fix purge / stop button on home.html
- Added purge and stop in pump controller
- Added blink functionality and time control in rgb controller
- Integration of rgb led with alcohol meter

+++ gpio-galileo-helper.js
@@ -84,7 +84,10 @@ GalileoGpio.prototype.openGPIO = function()
 GalileoGpio.prototype.openPin = function(pin_number)
 {
     //TODO: use rough GPIO method once available
-    fs.writeFileSync(this.gpioPath + "/export", this.pin[pin_number].id);
+    if (!fs.existsSync(this.gpioPath + "/gpio" + this.pin[pin_number].id))
+    {
+        fs.writeFileSync(this.gpioPath + "/export", this.pin[pin_number].id);
+    }
 };
 
 /**
@@ -167,7 +170,10 @@ GalileoGpio.prototype.readDigitalPin = function(pin_number)
  */
 GalileoGpio.prototype.closePin = function(pin_number)
 {
-    fs.writeFileSync(this.gpioPath + "/unexport", this.pin[pin_number].id);
+    if (fs.existsSync(this.gpioPath + "/gpio" + this.pin[pin_number].id))
+    {
+        fs.writeFileSync(this.gpioPath + "/unexport", this.pin[pin_number].id);
+    }
 };
 
 /**
